Added the placement of the CoreStamp to the planner
Additional removed the old visulization parts.
And start using the roomManager to call the planner. Also it gives the existing spawn as starting location.

The implementation of getting a start position is for now done. What both start positions share is getting the upgrader location and populate the room with the rest of the structures needed.