Minor refactor prepared the resource service
To be implemented with Minerals later on. Currently most of the energy providing side is done.

I will add a Porter for transporting energy to the terminal to sell, or move minerals to the terminal or labs. As well surplying upgrader and fast filler containers.

In this commit I also updated the Miners to use the container position to mine and set the objective spots size to 1.