Added scouting and scoutRoute creation
Initially the scout Route is created in the room initialization. After that it gets manipulated by the scouts to come.

Currently the scouts only append to the source info and ignore soure keeper rooms.

For remote mining this info should be enough, but I should catalouge more data that the room can give.

And avoid player rooms :D